squashed bugs and deleted a file fixed rendering bugs in the ledger and job ticket, if some fields were left blank the page wouldnt render
